add delete functionality, not tested

+++ modules/dfs/dfsmanager.py
@@ -150,13 +150,23 @@ class DFSManager:
 
     def delete_file(self, filename):
         ## remove from disk (if present)
-        self._filewriter.remove(filename)
         ## remove from _fs
-        self._filewriter.remove(filename) # does this need to check for existence of filename??
+        ## tell network to tell replicas
+        file = self._fs.get_file(filename)
+        if not file:
+            print("Invalid name")
+            return
+
+        # Check that you are a replica before removing file from disk
+        file_replicas = file["replicas"]
+        if self._id in file_replicas:
+            self._filewriter.remove(filename)
+
+        ## remove from _fs
         self._fs.delete_file(filename)
 
         self._network.delete_file(filename)
-        ## tell network to tell replicas        
+
 
     def set_priority(self, filename, priority):
         ## punt
